+++ 4 SEMESTRE/Pesquisa-Operacional-2/TrabalhoPesquisaOperacional/menorCusto.h
@@ -5,7 +5,7 @@
 
 infoProblemaTransporte achaMenorCusto(infoProblemaTransporte);
 
-int ** calculaSolucaoMenorCusto(infoProblemaTransporte);
+double ** calculaSolucaoMenorCusto(infoProblemaTransporte);
 
 void calculaSolucaComRotasReais(infoProblemaTransporte *);
